requirements.txt
pyrebase
python-telegram-bot
wikipedia
google_trans_new
urllib3
socket